Added Player Audio
Added Jack Footstep when walking using a raycast to each footstep which
is triggered by the animation. The code also allows for different ground
types e.g. water and grass. The ground is tagged and is checked to see
which audio clip to play. This chip is played at random volume.